requirements.txt
Flask
PyMuPDF
fpdf2
google-generativeai
python-dotenv